% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasystoreNGC-package.R
\docType{package}
\name{diseasystoreNGC-package}
\alias{diseasystoreNGC}
\alias{diseasystoreNGC-package}
\title{diseasystoreNGC: diseasystore extensions for the model group}
\description{
This package handles the feature store data model for the model group.
}
\author{
\strong{Maintainer}: Rasmus Skytte Randløv \email{rske@ssi.dk} (\href{https://orcid.org/0000-0002-5860-3838}{ORCID})

}
\keyword{internal}
