% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasystoreSarsCoV2.R
\name{DiseasystoreSarsCoV2}
\alias{DiseasystoreSarsCoV2}
\title{Feature store for individual level SARS-CoV-2}
\description{
This \code{DiseasystoreSarsCoV2} \link[R6:R6Class]{R6} brings support for individual level feature stores
}
\section{Super classes}{
\code{\link[diseasystore:DiseasystoreBase]{diseasystore::DiseasystoreBase}} -> \code{\link[diseasystoreNGC:DiseasystoreGeneric]{diseasystoreNGC::DiseasystoreGeneric}} -> \code{DiseasystoreSarsCoV2}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasystoreSarsCoV2-clone}{\code{DiseasystoreSarsCoV2$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="finalize"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-finalize'><code>diseasystore::DiseasystoreBase$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="get_feature"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-get_feature'><code>diseasystore::DiseasystoreBase$get_feature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="initialize"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-initialize'><code>diseasystore::DiseasystoreBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="key_join_features"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-key_join_features'><code>diseasystore::DiseasystoreBase$key_join_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="key_join_filter"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-key_join_filter'><code>diseasystore::DiseasystoreBase$key_join_filter()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreSarsCoV2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreSarsCoV2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreSarsCoV2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
